+++ lib/slimcop/cli.rb
@@ -1,9 +1,19 @@
 # frozen_string_literal: true
 
 require 'optparse'
+require 'rainbow'
 
 module Slimcop
   class Cli
+    COLOR_FOR_SEVERITY_CODE = {
+      convention: :yellow,
+      error: :red,
+      fatal: :red,
+      info: :gray,
+      refactor: :yellow,
+      warning: :magenta
+    }.freeze
+
     def initialize(argv)
       @argv = argv.dup
       @configuration = Configuration.new
@@ -52,12 +62,13 @@ module Slimcop
     def report(offenses)
       lines = offenses.map do |offense|
         format(
-          '%<file_path>s:%<line>i:%<column>i %<severity_code>s: %<message>s',
+          '%<file_path>s:%<line>i:%<column>i %<severity_code>s: %<correctability>s%<message>s',
           column: offense.real_column,
-          file_path: offense.file_path,
+          correctability: offense.correctable? ? Rainbow('[Correctable] ').yellow : '',
+          file_path: Rainbow(offense.file_path).cyan,
           line: offense.line,
           message: offense.message,
-          severity_code: offense.severity_code
+          severity_code: Rainbow(offense.severity.code).color(COLOR_FOR_SEVERITY_CODE[offense.severity.name])
         )
       end
       puts(lines)