+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Improve offenses CLI output format.
+- Return exit status 1 if any offense found.
 
 ## 0.1.0 - 2021-12-21