+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improve offenses CLI output format.
 - Return exit status 1 if any offense found.
 - Add severity code to offense output.
+- Colorize output.
 
 ## 0.1.0 - 2021-12-21